+++ jgclark.NoteHelpers/src/index.js
@@ -3,7 +3,7 @@
 // -----------------------------------------------------------------------------
 // Note Helpers plugin for NotePlan
 // Jonathan Clark & Eduard Metzger
-// last updated 22.3.2023 for v0.16.0, @jgclark
+// last updated 10.6.2023 for v0.17.0, @jgclark
 // -----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
@@ -22,7 +22,8 @@ export {
   jumpToNoteHeading,
   openCurrentNoteNewSplit,
   openNoteNewWindow,
-  openNoteNewSplit
+  openNoteNewSplit,
+  openURLFromANote
 } from './noteNavigation'
 export {
   addTriggerToNote,
@@ -31,6 +32,10 @@ export {
   moveNote,
   renameNoteFile,
 } from './noteHelpers'
+export {
+  // getDuplicateNotes,
+  showDuplicates,
+} from './duplicates'
 
 export function init(): void {
   // In the background, see if there is an update to the plugin to install, and if so let user know
@@ -71,4 +76,8 @@ export async function updateSettings() {
   } catch (error) {
     logError(pluginJson, JSP(error))
   }
+}
+
+export function resetCaches() {
+  NotePlan.resetCaches()
 }