+++ jgclark.SearchExtensions/src/flexiSearch.js
@@ -11,12 +11,24 @@ import { clo, logDebug, logError, logWarn } from '@helpers/dev'
 import { type HtmlWindowOptions, showHTMLV2 } from '@helpers/HTMLView'
 import { closeWindowFromCustomId, logWindowsList } from '@helpers/NPWindows'
 
+const pluginID = "jgclark.SearchExtensions"
 
 //-----------------------------------------------------------------------------
-const flexiSearchDialogHTML = ` <form type="dialog" id="searchOptions">
+const flexiSearchDialogHTML = `
+<div class="dialogBox">
+ <form type="dialog" id="searchOptions">
   <div class="dialogSection">
-		<b>Search Terms</b> <input type="text" id="searchTerms" name="searchTerms" style="width: 16rem;" value="#test" />
-    <br />(Separate terms by spaces. You can use +term, -term and !term as well.)
+		<b>Search Terms</b><input type="text" id="searchTerms" name="searchTerms" size="25" value="#test" autofocus tabindex="1" />
+    <br />
+    (Separate terms by spaces. You can use +term, -term and !term as well. See <a href="https://github.com/NotePlan/plugins/tree/main/jgclark.SearchExtensions/" target="_blank">README</a> for more details.)
+	</div>
+
+	<div class="dialogSection">
+		<b>Save results to </b>
+    <input type="radio" name="savetype" id="quick" value="quick" />
+    <label for="notetype">'Quick Search' note</label>
+    <input type="radio" name="savetype" id="newnote" value="newnote" />
+    <label for="notetype">Specific note</label>
   </div>
 
   <div class="dialogSection">
@@ -25,6 +37,8 @@ const flexiSearchDialogHTML = ` <form type="dialog" id="searchOptions">
     <label for="notetype">Regular notes</label>
     <input type="checkbox" name="notetype" id="calendar" value="calendar"  />
     <label for="notetype">Calendar notes</label>
+    <!-- following will normally be hidden by CSS -->
+    <span class="atLeastOneCheckboxWarning">[Please select at least one!]</span>
   </div>
 
   <div class="dialogSection">
@@ -40,24 +54,18 @@ const flexiSearchDialogHTML = ` <form type="dialog" id="searchOptions">
     <label for="task"><i class="fa-regular fa-circle"></i>Open</label>
 		</li>
 		<li>
-    <input type="checkbox" id="taskDone" name="task" value="done"  />
-    <label for="task"><i class="fa-regular fa-circle-check"></i>Complete</label>
-		</li>
-		<li>
     <input type="checkbox" id="taskScheduled" name="task"
       value="taskScheduled" />
     <label for="task"><i class="fa-regular fa-clock"></i>Scheduled</label>
 		</li>
 		<li>
+    <input type="checkbox" id="taskDone" name="task" value="done"  />
+    <label for="task"><i class="fa-regular fa-circle-check"></i>Complete</label>
+		</li>
+		<li>
     <input type="checkbox" id="taskCancelled" name="task" value="taskCancelled" />
     <label for="task"><i class="fa-regular fa-circle-xmark"></i>Cancelled</label>
 		</li>
-    <!--
-		<li>
-    <input type="checkbox" id="taskNone" name="task" value="" />
-    <label for="task">None</label>
-		</li>
-    -->
     </ul>
 	</td>
 	<td>
@@ -69,25 +77,18 @@ const flexiSearchDialogHTML = ` <form type="dialog" id="searchOptions">
     <label for="checklist"><i class="fa-regular fa-square"></i>Open</label>
 		</li>
 		<li>
-    <input type="checkbox" id="checklistDone" name="checklist" value="checklistDone" checked />
-    <label for="checklist"><i class="fa-regular fa-square-check"></i>Complete</label>
-		</li>
-		<li>
     <input type="checkbox" id="checklistScheduled" name="checklist"
       value="checklistScheduled" />
-    <!-- TODO: other icons needed -->
     <label for="checklist"><i class="fa-regular fa-square-chevron-right"></i>Scheduled</label>
+		</li>
+		<li>
+    <input type="checkbox" id="checklistDone" name="checklist" value="checklistDone" checked />
+    <label for="checklist"><i class="fa-regular fa-square-check"></i>Complete</label>
 		</li>
 		<li>
     <input type="checkbox" id="checklistCancelled" name="checklist" value="checklistCancelled" />
     <label for="checklist"><i class="fa-regular fa-square-xmark"></i>Cancelled</label>
 		</li>
-    <!--
-		<li>
-    <input type="checkbox" id="checklistNone" name="checklist" value="" />
-    <label for="checklist">None</label>
-		</li>
-    -->
   </ul>
 	</td>
 	<td>
@@ -95,15 +96,15 @@ const flexiSearchDialogHTML = ` <form type="dialog" id="searchOptions">
 		<li>Other line types:</li>
 		<li>
     <input type="checkbox" name="other" id="list" value="list" checked />
-    <label for="checklist"><kbd>-</kbd> Bullet lists</label>
+    <label for="checklist"><kbd>-</kbd>Bullet lists</label>
 		</li>
 		<li>
     <input type="checkbox" name="other" id="quote" value="quote" checked />
-    <label for="checklist"><kbd>&gt;</kbd> Quotations</label>
+    <label for="checklist"><kbd>&gt;</kbd>Quotations</label>
 		</li>
 		<li>
     <input type="checkbox" name="other" id="headings" value="title" checked />
-    <label for="checklist"><kbd>#</kbd> Headings</label>
+    <label for="checklist"><kbd>#</kbd>Headings</label>
 		</li>
 		<li>
     <input type="checkbox" name="other" id="text" value="text" checked />
@@ -115,17 +116,20 @@ const flexiSearchDialogHTML = ` <form type="dialog" id="searchOptions">
 	</table>
 
 	</div>
-  <div class="dialogSection buttonRow">
-    <input type="submit" value="Cancel" />
-    <input type="submit" value="Search" />
+  <div class="dialogSection">
+    <div class="buttonRow">
+    <input type="submit" value="Search" class="mainButton" tabindex="2"/>
+    <input type="submit" value="Cancel" id="displayFirst" tabindex="3"/>
+    </div>
   </div>
-</form>
+ </form>
+</div>
 `
 
 // Script to start the search options to the plugin and start it
 const JSStartSearchInPlugin = JSON.stringify(`
 (async function() {
-  await DataStore.invokePluginCommandByName('flexiSearchHandler', 'jgclark.SearchExtensions', ['%%SEARCHTERMS%%', '%%NOTETYPES%%', '%%PARATYPES%%'] )
+  await DataStore.invokePluginCommandByName('flexiSearchHandler', 'jgclark.SearchExtensions', ['%%SEARCHTERMS%%', '%%SAVETYPE%%', '%%NOTETYPES%%', '%%PARATYPES%%'] )
 })()
 `)
 
@@ -136,10 +140,10 @@ const JSCloseDialog = JSON.stringify(`
 })()
 `)
 
-// TODO: Script to save item to pref
+// Script to save item to DataStore.preference
 const JSUpdatePref = JSON.stringify(`
 (async function() {
-  await DataStore.invokePluginCommandByName('???', 'jgclark.SearchExtensions', ['%%KEY%%', '%%VALUE%%'] )
+  await DataStore.invokePluginCommandByName('savePluginPreference', 'jgclark.SearchExtensions', ['%%KEY%%', '%%VALUE%%'] )
 })()
 `)
 
@@ -147,45 +151,44 @@ const flexiSearchDialogPostBodyScripts = `
 <script type="text/javascript">
   window.addEventListener("load", () => {
     console.log('onLoad script running ...')
+
 		// Set defaults to use.
-		// Note following code assumes case sensitive matching, and that the values are distinct and not subset strings of each other
-		let noteTypesStr = 'notes,calendar,'
-		let paraTypesStr = 'open,done,checklistOpen,checklistDone,list,quote,title,text,'
+		// Note following code assumes case sensitive matching, and that the values are distinct and not subset strings of each other.
+    // Their values are substituted before the script is loaded
+    let saveType = '%%SAVETYPEPREF%%'
+		let noteTypesStr = '%%NOTETYPESSTRPREF%%'
+		let paraTypesStr = '%%PARATYPESSTRPREF%%'
     const formID = "searchOptions"
 		// Get the form element + input controls
     const form = document.getElementById(formID)
 		const inputs = form.elements
 
-		function loadDialogState() {
-			// TODO: load it or have it passed in some way
-			console.log('loadDialogState()')
-		}
-
-		loadDialogState()
-
 		// Iterate over checkbox controls setting whether they're initially checked or not
+    // Note additionaly complexity because 'list' is a substring of '...Checklist'
 		function initDialogState() {
 			console.log('initDialogState()')
+      const paraTypesArr = paraTypesStr.replace(/,{2,}/g, ',').replace(/,$/, '').replace(/^,/, '').split(',')
 			for (let i = 0; i < inputs.length; i++) {
-				if (inputs[i].type === "checkbox") {
-					const val = inputs[i].value
-					if (inputs[i].name === "notetype") {
-						console.log('- setting noteTypesStr "'+ val +'" to ' + noteTypesStr.includes(val))
-						inputs[i].checked = noteTypesStr.includes(val)
-					} else {
-						console.log('- setting paraTypesStr "'+ val +'" to ' + noteTypesStr.includes(val))
-						inputs[i].checked = paraTypesStr.includes(val)
-					}
-				}
+        const val = inputs[i].value
+        if (inputs[i].name === "notetype") {
+          console.log('- setting noteTypesStr "'+ val +'" to ' + String(noteTypesStr.includes(val)))
+          inputs[i].checked = noteTypesStr.includes(val)
+        } else if (inputs[i].name === "savetype") {
+          console.log('- setting saveType "'+ val +'" to ' + String(saveType === val))
+          inputs[i].checked = (saveType === val)
+        } else if (inputs[i].type === "checkbox") {
+          console.log('- setting paraTypesStr "'+ val +'" to ' + String(noteTypesStr.includes(val)))
+          inputs[i].checked = paraTypesArr.includes(val)
+        }
 			}
 		}
 
 		initDialogState()
 
 		function saveDialogState() {
+			console.log('saveDialogState()')
 			noteTypesStr = ''
 			paraTypesStr = ''
-			console.log('saveDialogState()')
 			// Iterate over the optional controls
 			for (let i = 0; i < inputs.length; i++) {
 				// console.log(inputs[i].nodeName, inputs[i].type, inputs[i].checked, inputs[i].value)
@@ -193,26 +196,33 @@ const flexiSearchDialogPostBodyScripts = `
 					// Add this checked value to a CSV string
 					noteTypesStr += inputs[i].value + ','
 				}
+				if (inputs[i].checked && (inputs[i].name === "savetype")) {
+					// Set this
+					saveType = inputs[i].value
+				}
 				if (inputs[i].checked && (inputs[i].name === "task" || inputs[i].name === "checklist" || inputs[i].name === "other")) {
 					// Add this checked value to a CSV string
 					paraTypesStr += inputs[i].value + ','
 				}
 			}
-      // TODO: warning if noteTypesStr is empty
-
-			// TODO: save
-			console.log('Saving ' + noteTypesStr + ' / ' + paraTypesStr)
+			console.log('Saving ' + saveType + ' / ' + noteTypesStr + ' / ' + paraTypesStr)
       window.webkit.messageHandlers.jsBridge.postMessage({
-          code: ${JSUpdatePref}.replace(%%KEY%%, 'noteTypesStr').replace(%%VALUE%%, noteTypesStr),
-          function: "miscFunc",
-          id: "1"
-        })
+        code: ${JSUpdatePref}.replace('%%KEY%%', 'saveType').replace('%%VALUE%%', saveType),
+        onHandle: "miscFunc",
+        id: "1"
+      })
       window.webkit.messageHandlers.jsBridge.postMessage({
-          code: ${JSUpdatePref}.replace(%%KEY%%, 'paraTypesStr').replace(%%VALUE%%, paraTypesStr),
-          function: "miscFunc",
-          id: "1"
-        })
-      }
+        code: ${JSUpdatePref}.replace('%%KEY%%', 'noteTypesStr').replace('%%VALUE%%', noteTypesStr),
+        onHandle: "miscFunc",
+        id: "1"
+      })
+      window.webkit.messageHandlers.jsBridge.postMessage({
+        code: ${JSUpdatePref}.replace('%%KEY%%', 'paraTypesStr').replace('%%VALUE%%', paraTypesStr),
+        onHandle: "miscFunc",
+        id: "1"
+      })
+      console.log('end of saveDialogState()')
+    }
 
     // Add 'change' event handler to form
 		form.addEventListener("change", (event) => {
@@ -223,7 +233,7 @@ const flexiSearchDialogPostBodyScripts = `
     form.addEventListener("submit", (event) => {
       event.preventDefault()
       const submitterValue = event.submitter.value
-      console.log('submit event fired with value '+submitterValue)
+      console.log('submit event fired with value ' + submitterValue)
 
       // Close if user has cancelled
       if (submitterValue === 'Cancel') {
@@ -231,7 +241,7 @@ const flexiSearchDialogPostBodyScripts = `
         // Note: can't just do 'window.close()' as the window wasn't open by a window.open() command
         window.webkit.messageHandlers.jsBridge.postMessage({
           code: ${JSCloseDialog},
-          function: "miscFunc",
+          onHandle: "miscFunc",
           id: "1"
         })
         return
@@ -239,28 +249,28 @@ const flexiSearchDialogPostBodyScripts = `
 
 			// Get text input
 			const searchTerms = inputs["searchTerms"].value
-			console.log(searchTerms)
 
       // Remove any multiple or leading or trailing comma(s)
       let noteTypes = noteTypesStr.replace(/,{2,}/g, ',').replace(/,$/, '').replace(/^,/, '')
       noteTypes = (noteTypes === 'notes,calendar') ? 'both' : noteTypes
-      // console.log(noteTypes)
       let paraTypes = paraTypesStr.replace(/,{2,}/g, ',').replace(/,$/, '').replace(/^,/, '')
-      // console.log(paraTypesStr)
-      // console.log(paraTypes)
 
       // Update the JS to send to the plugin based on the form values, and then send
       window.webkit.messageHandlers.jsBridge.postMessage({
-        code: ${JSStartSearchInPlugin}.replace('%%SEARCHTERMS%%', searchTerms).replace('%%NOTETYPES%%', noteTypes).replace('%%PARATYPES%%', paraTypes),
+        code: ${JSStartSearchInPlugin}
+          .replace('%%SEARCHTERMS%%', searchTerms)
+          .replace('%%SAVETYPE%%', saveType)
+          .replace('%%NOTETYPES%%', noteTypes)
+          .replace('%%PARATYPES%%', paraTypes),
         onHandle: "miscFunc",
         id: "1"
       })
     })
+  })
 
-    // placeholder function; not sure why it's needed, but it is!
-    function miscFunc(re, id) {
-    }
-  });
+  // placeholder function; not sure why it's needed, but it is!
+  function miscFunc(re, id) {
+  }
 
 </script>
 `
@@ -275,13 +285,27 @@ const headerTags = `
 
 const dialogCSS = `
   /* Speciifc CSS for the dialog box */
+  .dialogBox {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: var(--bg-main-color);
+    min-width: 20rem;
+    min-height: 17rem;
+  }
+	.dialogBox > div {
+    max-width: 25rem;
+    max-height: 22rem;
+    background-color: var(--bg-alt-color);
+  }
 	ul.dialogList {
 		list-style-type: none;
     margin: 0rem;
     padding-inline: 0.5rem;
 	}
   .dialogSection {
-    margin: 0.3rem 0rem;
+    margin: 0.4rem 0rem; /* horiz + vert */
   }
   .dialogSection b {
     font-weight: bold;
@@ -292,12 +316,31 @@ const dialogCSS = `
     padding-right: 0.4rem;
   }
   input[type="submit"] {
-    padding: 3px 6px 4px 6px;
     margin: 3px 6px;
-    font-size: 0.8rem;
+    padding: 3px 6px 4px 6px;
+    font-size: 0.9rem;
+  }
+  input[type="text"] {
+    margin-left: 0.4rem;
+    padding-left: 0.2rem;
+    font-size: 0.9rem;
   }
   .buttonRow {
-    text-align: center;
+    display: flex;
+    flex-direction: row-reverse;
+    justify-content: right;
+    align-items: right;
+  }
+	.mainButton {
+		font-weight: 700;
+	}
+  .atLeastOneCheckboxWarning {
+    font-size: 0.9rem;
+    color: red;
+    display: inline;
+  }
+  input:checked ~ span { /* when at least one of the checkboxes is checked its sibling span is hidden */
+    display: none;
   }
 `
 
@@ -309,9 +352,17 @@ const dialogCSS = `
 export async function flexiSearchRequest(
 ): Promise<void> {
   try {
+    // Look up the 3 preferences from local store
+    // Note: extra commas aren't typos
+    const saveType = String(DataStore.preference(pluginID + '.saveType')) ?? 'quick'
+    const noteTypesStr = String(DataStore.preference(pluginID + '.noteTypesStr')) ?? 'notes,calendar,'
+    const paraTypesStr = String(DataStore.preference(pluginID + '.paraTypesStr')) ?? 'open,done,checklistOpen,checklistDone,list,quote,title,text,'
+    const flexiSearchDialogPostBodyScriptsWithPrefValues = flexiSearchDialogPostBodyScripts
+      .replace('%%SAVETYPEPREF%%', saveType)
+      .replace('%%NOTETYPESSTRPREF%%', noteTypesStr)
+      .replace('%%PARATYPESSTRPREF%%', paraTypesStr)
 
     logDebug(pluginJson, `flexiSearchRequest called`)
-    // TODO: add appropriate FA libraries
     // write HTML to capture relevant search options
     const opts: HtmlWindowOptions = {
       windowTitle: 'FlexiSearch',
@@ -319,50 +370,100 @@ export async function flexiSearchRequest(
       headerTags: headerTags,
       generalCSSIn: '',
       specificCSS: dialogCSS,
-      makeModal: false, // TODO: is this more helpful?
-      postBodyScript: flexiSearchDialogPostBodyScripts,
+      makeModal: false, // modal doesn't actually help us here
+      postBodyScript: flexiSearchDialogPostBodyScriptsWithPrefValues,
       savedFilename: '../../jgclark.SearchExtensions/flexiSearchDialog.html',
-      width: 400, // FIXME: being ignored
-      height: 300,
-      reuseUsersWindowRect: false,
-      shouldFocus: true, // FIXME: being ignored?
+      width: 440,
+      height: 350,
+      reuseUsersWindowRect: true,
+      shouldFocus: true,
     }
     // show dialog as non-modal HTML window
     await showHTMLV2(flexiSearchDialogHTML, opts)
   }
   catch (err) {
-    logError(pluginJson, err.message)
+    logError(pluginJson, 'flexiSearcRequest: ' + err.message)
   }
 }
 
-export function flexiSearchHandler(searchTerms: string, noteTypeToInclude: string, paraTypes: string): any {
+export function flexiSearchHandler(searchTerms: string, saveType: string, noteTypeToInclude: string, paraTypes: string): any {
   try {
-    logDebug(pluginJson, `flexiSearchHandler called with ${searchTerms} / ${noteTypeToInclude} / ${paraTypes}`)
+    logDebug(pluginJson, `flexiSearchHandler called with [${searchTerms}] / ${saveType} / ${noteTypeToInclude} / ${paraTypes}`)
     // First close the window
     closeDialogWindow('flexiSearchDialog')
 
+    // Take saveType and noteTypeToInclude add create originatorCommand from it
+    let originatorCommand =
+      (saveType === 'quick') ? 'quickSearch'
+        : (noteTypeToInclude === 'notes') ? 'searchOverNotes'
+          : (noteTypeToInclude === 'calendar') ? 'searchOverCalendar'
+            : 'search' // which defaults to 'both'
+
     // Then call main saveSearch function (no need to await for it)
-    // FIXME: how to deal with case of empty search terms going to previous dialog?
-    saveSearch(searchTerms, noteTypeToInclude, 'newnote', paraTypes, 'Searching')
+    // TODO: how to deal with case of empty search terms going to previous dialog?
+    saveSearch(searchTerms, noteTypeToInclude, originatorCommand, paraTypes, 'Searching')
     return {} // apparently required to avoid error in log
   }
   catch (err) {
-    logError(pluginJson, err.message)
+    logError(pluginJson, 'flexiSearchHandler: ' + err.message)
     return {}
   }
 }
 
+/**
+ * Way for an HTML window to request that it be closed.
+ * Is there a simpler way? I can't find one yet.
+ * @param {customId} customId
+ * @returns {any} not used, but has to be present
+ */
 export function closeDialogWindow(customId: string): any {
   try {
-    logDebug(pluginJson, `closeDialogWindow('${customId}') called`)
-    logWindowsList()
-    // Close the window
+    // logDebug(pluginJson, `closeDialogWindow('${customId}') called`)
     closeWindowFromCustomId(customId)
 
     return {} // apparently required to avoid error in log
   }
   catch (err) {
-    logError(pluginJson, err.message)
+    logError(pluginJson, 'closeDialogWindow: ' + err.message)
+    return {}
+  }
+}
+
+/**
+ * Helper function for HTML views to set a DataStore.preference value (as a string)
+ * @param {string} key to set
+ * @param {string} value to set
+ * @returns {any}
+ */
+export function savePluginPreference(key: string, value: string): any {
+  try {
+    logDebug(pluginJson, `savePluginPreference('${key}', '${value}') called for ${pluginID}`)
+    DataStore.setPreference(pluginID + '.' + key, value)
+    // logDebug(pluginJson, `-> ${DataStore.preference(pluginID + '.' + key)}}`)
+
+    return {} // apparently required to avoid error in log
+  }
+  catch (err) {
+    logError(pluginJson, 'savePluginPreference: ' + err.message)
+    return {}
+  }
+}
+
+/**
+ * Helper function for HTML views to get DataStore.preference value
+ * @param {string} key to read
+ * @returns {any}
+ */
+export function getPluginPreference(key: string): any {
+  try {
+    logDebug(pluginJson, `getPluginPreference('${key}') called for ${pluginID}`)
+    // logDebug(pluginJson, `-> ${DataStore.preference(pluginID + '.' + key)}}`)
+    return DataStore.preference(pluginID + '.' + key)
+
+    // return {} // apparently required to avoid error in log
+  }
+  catch (err) {
+    logError(pluginJson, 'getPluginPreference: ' + err.message)
     return {}
   }
 }